Divided the code into three pages
